Break instrumentation out by adapter
